[#193] feat : settings 페이지 서버렌더링
