+++ app/settings/_components/MyProfile/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import * as styles from "./style.css";
 import CheckRigntIcon from "@/public/icons/chevron-right.svg?url";
 import Image from "next/image";
@@ -7,18 +9,17 @@ import { getMe } from "@/app/_api/users";
 import { getImagePath } from "@/app/_utils/getPersonImagePath";
 
 const MyProfile = () => {
-  const { data: user, isLoading } = useQuery<UserType>({
+  const { data } = useQuery<UserType>({
     queryKey: ["me"],
     queryFn: () => getMe(),
   });
 
-  if (isLoading) return <div>Loading...</div>;
-  const profileImageUrl = user && getImagePath(user.profilePath);
+  const profileImageUrl = data && getImagePath(data.profilePath);
 
   return (
     <div className={styles.container}>
       <div className={styles.title}>마이프로필 관리하기</div>
-      {user && (
+      {data && (
         <div className={styles.ProfileWrapper}>
           <div
             className={styles.profileImg}
@@ -27,8 +28,8 @@ const MyProfile = () => {
             }}
           />
           <div>
-            <p className={styles.nickname}>{user.nickname}</p>
-            <p className={styles.email}>{user.email}</p>
+            <p className={styles.nickname}>{data.nickname}</p>
+            <p className={styles.email}>{data.email}</p>
           </div>
           <Image className={styles.checkRightIcon} src={CheckRigntIcon} alt="check right icon" width={22} height={22} />
         </div>
